% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bin_cv_loglik}
\alias{bin_cv_loglik}
\title{bin_cv_loglik cpp}
\usage{
bin_cv_loglik(x, beta, y, time)
}
\arguments{
\item{x}{design matrix}

\item{beta}{vcm vector}

\item{y}{response vector}

\item{time}{vector}
}
\description{
LogLikilihood function for cv
}
