% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bin_logit_loglik}
\alias{bin_logit_loglik}
\title{bin_logit_loglik cpp}
\usage{
bin_logit_loglik(x, beta, y, time, time_zero, h, type)
}
\arguments{
\item{x}{design matrix}

\item{beta}{vcm vector}

\item{y}{response vector}

\item{time}{vector}

\item{time_zero}{grid point double}

\item{h}{bandwidth double}

\item{type}{kernel function int}
}
\description{
LogLikilihood function for vcm
}
