% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binary_data}
\alias{binary_data}
\title{A simulated dataset of 200 observations with each containing 25 correlated binary response variables.}
\format{
A data frame with 5000 rows and 5 variables:
\describe{
  \item{id}{A variable grouping the repeated measurements}
  \item{time}{The time for each repeated measurement}
  \item{x1}{First predictor variable}
  \item{x2}{Second predictor variable}
  \item{Y}{Correlated binary response variable}
 }
}
\usage{
binary_data
}
\description{
A simulated dataset of 200 observations with each containing 25 correlated binary response variables.
}
\keyword{datasets}
