% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernel_function}
\alias{kernel_function}
\title{kernel_function}
\usage{
kernel_function(time, time_zero, h, type)
}
\arguments{
\item{time}{time point}

\item{time_zero}{grid point}

\item{h}{bandwidth}

\item{type}{kernel function}
}
\description{
Computes the weight using a kernel function
}
