% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernel_vec}
\alias{kernel_vec}
\title{kernel_vec}
\usage{
kernel_vec(time, time_zero, h, type)
}
\arguments{
\item{time}{time point as a vector}

\item{time_zero}{grid point}

\item{h}{bandwidth}

\item{type}{kernel function}
}
\description{
Vectorized kernel function
}
