% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pois_cv_loglik}
\alias{pois_cv_loglik}
\title{pois_cv_loglik cpp}
\usage{
pois_cv_loglik(x, beta, y, time)
}
\arguments{
\item{x}{design matrix}

\item{beta}{vcm vector}

\item{y}{response vector}

\item{time}{vector}
}
\description{
LogLikelihood function for cv
}
