% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_vcm_cv.R
\name{pois_vcm_cv}
\alias{pois_vcm_cv}
\title{pois_vcm_cv: Cross-Validation for Poisson Time-Varying Coefficient Model}
\usage{
pois_vcm_cv(
  formula,
  data,
  time,
  id = NULL,
  bandwidths = NULL,
  kernel = "epanechnikov"
)
}
\arguments{
\item{formula}{A formula class object. Provides information about the names of the response and predictor variables.}

\item{data}{A data frame containing the variables needed for the function.}

\item{time}{The name of the variable indicating the time points for each observation.}

\item{id}{The name of the variable to group observations. Necessary when dealing with repeated measurements such as longitudinal data.}

\item{bandwidths}{A vector containing the bandwidth values for the cross-validation. Default is tests 50 bandwidths from ranging from double the smallest time point difference to range of time points.}

\item{kernel}{The name of the kernel function used to estimate the varying-coefficient values. Default is Epanechnikov.}
}
\value{
pois_vcm_cv returns a list containing the results from the cross-validation
\itemize{
\item recommended_bandwidth: The bandwidth with the smallest negative log-likelihood.
\item value: value of negative log-likelihood at recommended bandwidth
\item neg_logliks: The tested bandwidths and their corresponding negative log-likelihood value.
}
}
\description{
This function conducts a leave-one-out cross-validation approach to determine the best bandwidth for the data.
}
\examples{

cv_fit <- pois_vcm_cv(formula = Y~x1+x2, data = pois_data, time = time, id = id)

summary(cv_fit)

}
\author{
Isaac Quintanilla Salinas
}
