% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tvcm.cv.R
\name{summary.tvcm.cv}
\alias{summary.tvcm.cv}
\title{summary.tvcm.cv}
\usage{
\method{summary}{tvcm.cv}(object, ...)
}
\arguments{
\item{object}{An object of class "tvcm.cv"}

\item{...}{further arguments; not in use}
}
\description{
Prints the recommended bandwidth.
}
