% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tvcm.R
\name{summary.tvcm}
\alias{summary.tvcm}
\title{summary.tvcm: Plots time-varying coefficient models}
\usage{
\method{summary}{tvcm}(object, bounds = FALSE, deriv = FALSE, ...)
}
\arguments{
\item{object}{An object of class "tvcm"}

\item{bounds}{If TRUE, the bootstrap percentiles will be plotted}

\item{deriv}{If TRUE, the derivatives will be plotted}

\item{...}{further arguments; not in use}
}
\value{
Plots
}
\description{
This function allows you to plot the results from any of time-varying coefficient functions.
}
\examples{

fit <- tvcm(formula = Y~x1+x2, data = normal_data, time = time,
            id = id, se = TRUE, nboot = 100)
            
           
summary.tvcm(fit, bounds = TRUE, deriv = TRUE)


}
