% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvcm.R
\name{tvcm}
\alias{tvcm}
\title{tvcm: Time-Varying Coefficient Models}
\usage{
tvcm(
  formula,
  data,
  time,
  id = NULL,
  ngrids = 200,
  grid_points = NULL,
  bandwidth = NULL,
  kernel = "epanechnikov",
  se = FALSE,
  alpha = 0.05,
  nboot = 1000
)
}
\arguments{
\item{formula}{A formula class object. Provides information about the names of the response and predictor variables.}

\item{data}{A data frame containing the variables needed for the function.}

\item{time}{The name of the variable indicating the time points for each observation.}

\item{id}{The name of the variable to group observations. Necessary when dealing with repeated measurements such as longitudinal data.}

\item{ngrids}{When specified, a vector of size ngrid is created for the varying-coefficient values. The default value is 200. The vector is creates using the maximum and minimum from the provided time points.}

\item{grid_points}{A vector indicating the grid points to estimate the varying coefficient values. When specified, ngrid is ignored.}

\item{bandwidth}{A numeric value indicating the bandwidth. Default is the tenth of the range of the time points.}

\item{kernel}{The name of the kernel function used to estimate the varying-coefficient values. Default is Epanechnikov.}

\item{se}{If set TRUE, a bootstrap method is applied to estimate the standard errors and percentiles.}

\item{alpha}{A value indicating the significance level for the percentiles.}

\item{nboot}{A number indicating how many boot samples to construct.}
}
\value{
tvcm returns a list containing the estimated varying coefficients
\itemize{
\item estimates: A list containing two matrices:
\itemize{
\item est: a matrix for the estimates, each row represents the estimate of the varying coefficient
\item deriv: a matrix for the first derivative, each row represents the derivative of the varying coefficient
}
\item bootstrap_results: If specified, a list containing the elements below
\itemize{
\item boot_se: A list containing two matrices for the standard errors:
\itemize{
\item est: a matrix for the estimates, each row represents the estimate of the varying coefficient
\item deriv: a matrix for the first derivative, each row represents the derivative of the varying coefficient
}
\item boot_lower: A list containing two matrices for the lower percentiles:
\itemize{
\item est: a matrix for the estimates, each row represents the estimate of the varying coefficient
\item deriv: a matrix for the first derivative, each row represents the derivative of the varying coefficient
}
\item boot_upper:A list containing two matrices for the upper percentiles:
 \itemize{
\item est: a matrix for the estimates, each row represents the estimate of the varying coefficient
\item deriv: a matrix for the first derivative, each row represents the derivative of the varying coefficient
}
\item boot_samples: An array containing the bootstrap samples
}
\item time_points: The time points used for the estimates
}
}
\description{
This function estimates the time-varying coefficient model at different time points.
}
\examples{

tvcm(formula = Y~x1+x2, data = normal_data, time = time, 
     id = id, se = TRUE, nboot = 100)

}
\author{
Isaac Quintanilla Salinas
}
