% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tvcm_wls}
\alias{tvcm_wls}
\title{tvcm_wls cpp}
\usage{
tvcm_wls(x, y, time, id, time_zero, h, type)
}
\arguments{
\item{x}{design matrix}

\item{y}{response vector}

\item{time}{vector}

\item{id}{grouping vector}

\item{time_zero}{grid point double}

\item{h}{bandwidth double}

\item{type}{kernel function int}
}
\description{
Weighted Least-Squares Function For Longitudinal Data
}
