% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vcm_wls}
\alias{vcm_wls}
\title{vcm_wls cpp}
\usage{
vcm_wls(x, y, time, time_zero, h, type)
}
\arguments{
\item{x}{design matrix}

\item{y}{response vector}

\item{time}{vector}

\item{time_zero}{grid point double}

\item{h}{bandwidth double}

\item{type}{kernel function int}
}
\description{
Weighted Least Squares Function
}
